+++ db/migrate/20161110163819_remove_options_from_designs.rb
@@ -1,0 +1,5 @@
+class RemoveOptionsFromDesigns < ActiveRecord::Migration[5.0]
+  def change
+    remove_column :designs, :options, :text
+  end
+end